requirements.txt
openai
pypdf
python-dotenv
groq